Upgrade Tipue Search to v5.0
 - New JS and CSS files for Tipue Search v5.0.
 - LateTask to create "search.html" page to use for search entries
