--- v7/localsearch/localsearch/files/assets/css/tipuesearch.css
@@ -1,159 +1,0 @@
-
-/*
-Tipue Search 3.0.1
-Copyright (c) 2013 Tipue
-Tipue Search is released under the MIT License
-http://www.tipue.com/search
-*/
-
-
-#tipue_search_input
-{
-     font: 12px/1.7 'open sans', sans-serif;
-	color: #333;
-	padding: 7px;
-	width: 150px;
-	border: 1px solid #e2e2e2;
-	border-radius: 0;
-	-moz-appearance: none;
-	-webkit-appearance: none;
-     box-shadow: none;
-	outline: 0;
-	margin: 0;
-}
-#tipue_search_input:focus
-{
-     border: 1px solid #ccc;
-}
-#tipue_search_button
-{
-     width: 70px;
-     height: 36px;
-     border: 0;
-     border-radius: 1px;
-     background: #5193fb url('img/search.png') no-repeat center;
-	outline: none;
-}
-#tipue_search_button:hover
-{
-     background-color: #4589fb;
-}
-
-#tipue_search_content
-{
-     clear: left;
-     max-width: 650px;
-	padding: 25px 0 13px 0;
-	margin: 0;
-}
-#tipue_search_loading
-{
-     padding-top: 60px;
-     background: #fff url('img/loader.gif') no-repeat left;
-}
-
-#tipue_search_warning_head
-{
-     font: 300 16px/1.6 'open sans', sans-serif;
-	color: #333;
-}
-#tipue_search_warning
-{
-     font: 12px/1.6 'open sans', sans-serif;
-	color: #333;
-	margin: 7px 0;
-}
-#tipue_search_warning a
-{
-     color: #3f72d8;
-     text-decoration: none;
-}
-#tipue_search_warning a:hover
-{
-	padding-bottom: 1px;
-	border-bottom: 1px solid #ccc;
-}
-#tipue_search_results_count
-{
-     font: 13px/1.6 'open sans', sans-serif;
-	color: #333;
-}
-.tipue_search_content_title
-{
-     font: 300 23px/1.6 'open sans', sans-serif;
-	margin-top: 31px;
-}
-.tipue_search_content_title a
-{
-     color: #3f72d8;
-     text-decoration: none;
-}
-.tipue_search_content_title a:hover
-{
-	padding-bottom: 1px;
-	border-bottom: 1px solid #ccc;
-}
-.tipue_search_content_text
-{
-     font: 12px/1.7 'open sans', sans-serif;
-	color: #333;
-	padding: 13px 0;
-}
-.tipue_search_content_loc
-{
-     font: 300 13px/1.7 'open sans', sans-serif;
-     overflow: auto;
-}
-.tipue_search_content_loc a
-{
-     color: #555;
-     text-decoration: none;
-}
-.tipue_search_content_loc a:hover
-{
-	padding-bottom: 1px;
-	border-bottom: 1px solid #ccc;
-}
-#tipue_search_foot
-{
-     margin: 51px 0 21px 0;
-}
-#tipue_search_foot_boxes
-{
-	padding: 0;
-	margin: 0;
-     font: 12px/1 'open sans', sans-serif;
-}
-#tipue_search_foot_boxes li
-{
-	list-style: none;
-	margin: 0;
-	padding: 0;
-	display: inline;
-}
-#tipue_search_foot_boxes li a
-{
-	padding: 7px 13px 8px 13px;
-     background-color: #f1f1f1;
-     border: 1px solid #dcdcdc;
-     border-radius: 1px;
-	color: #333;
-	margin-right: 7px;
-	text-decoration: none;
-	text-align: center;
-}
-#tipue_search_foot_boxes li.current
-{
-	padding: 7px 13px 8px 13px;
-     background: #fff;
-     border: 1px solid #dcdcdc;
-     border-radius: 1px;
-	color: #333;
-	margin-right: 7px;
-	text-align: center;
-}
-#tipue_search_foot_boxes li a:hover
-{
-	border: 1px solid #ccc;
-	background-color: #f3f3f3;
-}
