--- v7/localsearch/localsearch/files/assets/js/tipuesearch_set.js
@@ -1,21 +1,0 @@
-
-/*
-Tipue Search 3.0.1
-Copyright (c) 2013 Tipue
-Tipue Search is released under the MIT License
-http://www.tipue.com/search
-*/
-
-
-var tipuesearch_stop_words = ["and", "be", "by", "do", "for", "he", "how", "if", "is", "it", "my", "not", "of", "or", "the", "to", "up", "what", "when"];
-
-var tipuesearch_replace = {"words": [
-     {"word": "tipua", replace_with: "tipue"},
-     {"word": "javscript", replace_with: "javascript"}
-]};
-
-var tipuesearch_stem = {"words": [
-     {"word": "e-mail", stem: "email"},
-     {"word": "javascript", stem: "script"},
-     {"word": "javascript", stem: "js"}
-]};
