--- v7/localsearch/localsearch/files/tipue_search.html
@@ -1,30 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-<title>Tipue Search</title>
-<meta charset="utf-8">
-
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.8.2/jquery.min.js"></script>
-
-<link rel="stylesheet" type="text/css" href="assets/css/tipuesearch.css">
-<script  src="assets/js/tipuesearch_set.js"></script>
-<scrip tsrc="assets/js/tipuesearch.js"></script>
-
-</head>
-<body>
-
-<div style="float: left;"><input type="text" id="tipue_search_input"></div>
-<div style="float: left; margin-left: 13px;"><input type="button" id="tipue_search_button"></div>
-<div id="tipue_search_content"><div id="tipue_search_loading"></div></div>
-</div>
-
-<script>
-$(document).ready(function() {
-     $('#tipue_search_input').tipuesearch({
-          'mode': 'json',
-          'contentLocation': 'assets/js/tipuesearch_content.json'
-     });
-});
-</script>
-</body>
-</html>
